+++ CHANGELOG.md
@@ -1,3 +1,14 @@
+### 2015/3/19
+
+- (categories/*.html を articles/<category>/index.html に：やめた)
+- ../archives を archives に（articles/ 直下に）
+- archive_yearly
+
+
+### 2015/3/18
+
+- stylesheet を cscc/compass で
+
 
 ### 2015/3/16